Phase-2/book-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
pydantic-extra-types==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0 